requirements.txt
# Playwright Automation with Ollama - Python Dependencies
# Updated to use Ollama instead of OpenAI

# Core dependencies
playwright==1.40.0
requests==2.31.0
python-dotenv==1.0.0

# CLI and UI
click==8.1.7
colorama==0.4.6

# Web scraping and parsing
beautifulsoup4==4.12.2

# Testing (optional)
pytest==7.4.3
pytest-playwright==0.4.3

# Removed OpenAI dependency - we now use Ollama via HTTP requests
# openai==1.3.0  # REMOVED - replaced with direct HTTP requests to Ollama

# Note: Make sure you have Ollama installed and running:
# 1. Install Ollama: curl -fsSL https://ollama.ai/install.sh | sh
# 2. Start Ollama: ollama serve
# 3. Pull a model: ollama pull llama2